requirements.txt
PyQt6>=6.6.0
PyQt6-Svg>=6.6.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
llama-cpp-python>=0.2.0
huggingface_hub>=0.20.0
